# Speed up the tokentx query